+++ src/game/actions/gathering/mining.js
@@ -2,9 +2,9 @@ const {initGathering, validateGathering} = require('./gatheringAction');
 
 const skillName = 'mining';
 
-async function validate(character, args, activeTimeout) {
+async function validate(character, args) {
   return new Promise(async (resolve, reject) => {
-    validateGathering(skillName, character, args, activeTimeout, resolve, reject);
+    validateGathering(skillName, character, args, resolve, reject);
   });
 }
