+++ src/game/actions/refining/woodworking.js
@@ -1,62 +1,17 @@
-const {Globals,CharacterService, validateLevel, validateIngredients, getRecipe, crafting } = require('../actionUtils')
+const {initCrafting, validateCrafting} = require('./craftingAction');
 
-const skillName = 'woodworking'
+const skillName = 'woodworking';
 
-async function startWoodworking(character, args, activeTimeout) {
-	return new Promise(async (resolve, reject) => {
-		console.log(`init Validation ${skillName}...`)
-
-		const recipeName = args.recipe
-		const recipe = getRecipe(skillName, recipeName)
-    if(!recipe) {
-			console.log(`${recipeName} does not exist for ${skillName}`)
-      reject('recipe')
-      return
-    }
-
-		const characterSkill = await CharacterService.getSkill(character, skillName);
-		const selectedResources = new Set()
-		//check is required level to use and used the recipe ingredients correctly
-		try {
-      await validateLevel(character, characterSkill.level, recipe.level)
-			console.log('validateLevel successfully')
-			await validateIngredients(character, args.ingredients, recipe, selectedResources)
-			console.log('validateIngredients successfully')
-    } catch (error) {
-			console.log('Validation failed: ', error.message)
-      reject(error.message);
-      return;
-    }
-		
-
-	// all the selected items are valid
-	console.log(`Validation complete. Init timeout ${skillName}...`)
-
-	let actionTime = Math.floor(recipe.time / (1 + characterSkill.speed));
-	const timeoutID = setTimeout(async () => {
-		// after the delay we craft!
-		try {
-			await crafting(character,skillName, recipeName, selectedResources)
-		} catch (error) {
-			console.log('crafting failed: ', error.message)
-      reject(error.message);
-      return;
-		}
-		activeTimeout[character] = null
-		resolve('success!')
-	}, Globals.getSpeedModifier()*actionTime)
-
-	console.log(`Init timeout with ${actionTime}ms complete. Waiting for crafting ${skillName}...`);
-
-	// setting a function to cancel the timeout
-	function cancelTimeout() {
-		console.log(`cancel timeout ${skillName}...`)
-		clearTimeout(timeoutID)
-		reject('cancel')
-	}
-	activeTimeout[character] = cancelTimeout
-
-	});
+async function validate(character, args) {
+  return new Promise(async (resolve, reject) => {
+    validateCrafting(skillName, character, args, resolve, reject);
+  });
 }
 
-module.exports = {startWoodworking}
+async function start(character, args, activeTimeout, actionTime) {
+  return new Promise(async (resolve, reject) => {
+    initCrafting(skillName, character, args, activeTimeout, resolve, reject, actionTime);
+  });
+}
+
+module.exports = {validate, start};
